requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# AI/LLM dependencies
langchain==0.1.0
langchain-anthropic==0.1.1

# Future SpoonOS dependencies (add as needed)
# web3==6.11.0
# requests==2.31.0